requirements.txt
# Google Generative AI SDK
google-genai>=0.3.0
google-generativeai>=0.8.0

# Groq (OPCIONAL - Alternativa ULTRA RÁPIDA ⚡)
groq>=0.4.0

# LangChain para RAG
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0

# Vector Store
chromadb>=0.4.22

# Reranker
cohere>=5.0.0

# PDF Processing
pypdf>=4.0.0
PyMuPDF>=1.23.0  # Para extração de imagens reais
Pillow>=10.0.0

# Interface
rich>=13.7.0
streamlit>=1.38.0

# Utils
python-dotenv>=1.0.0
tiktoken>=0.5.0

